+++ src/backend/process.py
@@ -1924,6 +1924,7 @@ class ProcessBackEnd:
             user_tokens=user_tokens,
             override_title_rules=override_title_rules,
             video_dynamic_range=self.config.settings.global_management.video_dynamic_range,
+            flat_filters=context.flat_filters,
             **self._release_info_token_kwargs(
                 release_info,
                 self.config.settings.series.multi_episode_style,